refactor: rename /src/utils to /src/aoc -- set of aoc related libs and helper functions, #44
