+++ src/2024/2024-12-05/lib/fileReader.ts
@@ -1,6 +1,6 @@
 import path from 'path'
-import { directory, readFile } from '@/utils/file.js'
-import { uniformArrayElements } from '@/utils/arrays.js'
+import { directory, readFile } from '@/aoc/file.js'
+import { uniformArrayElements } from '@/aoc/arrays.js'
 
 export type Rules = Record<number, number[]>
