+++ src/2024/2024-12-06/sample.test.ts
@@ -1,8 +1,8 @@
 import path from 'path'
 import { test, expect } from 'vitest'
 
-import { readAOCInputFile, AOC_OUTPUT_TYPE } from '@/utils/aocInputFile.js'
-import { directory } from '@/utils/file.js'
+import { readAOCInputFile, AOC_OUTPUT_TYPE } from '@/aoc/aocInputFile.js'
+import { directory } from '@/aoc/file.js'
 
 import { guardController } from './lib/guardController.js'
 import { findObstructionPositions } from './lib/guardControllerLoop.js'
