+++ src/2024/2024-12-09/sample.test.ts
@@ -1,8 +1,8 @@
 import path from 'path'
 import { test, expect } from 'vitest'
 
-import { AOC_OUTPUT_TYPE, readAOCInputFile } from '@/utils/aocInputFile.js'
-import { directory } from '@/utils/file.js'
+import { AOC_OUTPUT_TYPE, readAOCInputFile } from '@/aoc/aocInputFile.js'
+import { directory } from '@/aoc/file.js'
 
 import { CompactDisk } from './lib/compact.js'
 import { WholeDisk } from './lib/whole.js'
