+++ src/2024/2024-12-10/sample.test.ts
@@ -1,7 +1,7 @@
 import { test, expect } from 'vitest'
 
-import { AOC_OUTPUT_TYPE, readAOCInputFile } from '@/utils/aocInputFile.js'
-import { file } from '@/utils/file.js'
+import { AOC_OUTPUT_TYPE, readAOCInputFile } from '@/aoc/aocInputFile.js'
+import { file } from '@/aoc/file.js'
 
 import { countTrailScores } from './lib/scoresRatings.js'
