+++ src/2024/2024-12-12/main.ts
@@ -1,5 +1,5 @@
-import { AOC_OUTPUT_TYPE, readAOCInputFile } from '@/utils/aocInputFile.js'
-import { file } from '@/utils/file.js'
+import { AOC_OUTPUT_TYPE, readAOCInputFile } from '@/aoc/aocInputFile.js'
+import { file } from '@/aoc/file.js'
 
 import { Garden } from './lib/garden.js'
 import { WholesaleGarden } from './lib/wholesale.js'
