+++ src/aoc/aocInputFile.ts
