+++ src/aoc/arrays.ts
