+++ src/aoc/grid/utils.ts
