+++ src/aoc/numbers.ts
